requirements.txt
annotated-types==0.5.0 ; python_version >= "3.8" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.8" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.1.2 ; python_version >= "3.8" and python_version < "3.11"
fastapi==0.100.1 ; python_version >= "3.8" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.8" and python_version < "4.0"
google-auth==2.22.0 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-compute==1.13.0 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-monitoring==2.15.1 ; python_version >= "3.8" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.8" and python_version < "4.0"
grpcio-status==1.56.2 ; python_version >= "3.8" and python_version < "4.0"
grpcio==1.56.2 ; python_version >= "3.8" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.8" and python_version < "4.0"
protobuf==4.23.4 ; python_version >= "3.8" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.8" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic-core==2.4.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic==2.1.1 ; python_version >= "3.8" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.8" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.8" and python_version < "4"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.8" and python_version < "4.0"
uvicorn==0.23.2 ; python_version >= "3.8" and python_version < "4.0"
